Module openapi::apis::entries_api

Structs

LogEntries

Enums

CreateLogEntryError

struct for typed errors of method create_log_entry

GetLogEntryByIndexError

struct for typed errors of method get_log_entry_by_index

GetLogEntryByUuidError

struct for typed errors of method get_log_entry_by_uuid

SearchLogQueryError

struct for typed errors of method search_log_query

Functions

create_log_entry

Creates an entry in the transparency log for a detached signature, public key, and content. Items can be included in the request or fetched by the server when URLs are specified.

get_log_entry_by_index

get_log_entry_by_uuid

Returns the entry, root hash, tree size, and a list of hashes that can be used to calculate proof of an entry being included in the transparency log

search_log_query