Module openapi::v3_0

Support for OpenApi version 3.0.1 specification.

See the specification for more information.

Structs

AuthorizationCodeFlow
Configuration details for a authorization code OAuth Flow See [link] [link](https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md#oauth-flow-object

Callback
A map of possible out-of band callbacks related to the parent operation. Each value in the map is a Path Item Object that describes a set of requests that may be initiated by the API provider and the expected responses. The key value used to identify the callback object is an expression, evaluated at runtime, that identifies a URL to use for the callback operation.

ClientCredentialsFlow
Configuration details for a client credentials OAuth Flow See [link] [link](https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md#oauth-flow-object

Components
Holds a set of reusable objects for different aspects of the OAS.

Contact
Contact information for the exposed API.

Encoding
A single encoding definition applied to a single schema property.

Example
See https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md#exampleObject.

Extensions
Contains openapi specification extensions see https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md#specificationExtensions

ExternalDoc
Allows referencing an external resource for extended documentation.

Flows
Allows configuration of the supported OAuth Flows. See [link] [link][https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md#oauth-flows-object]

Header
The Header Object follows the structure of the Parameter Object with the following changes:

ImplicitFlow
Configuration details for a implicit OAuth Flow See [link] link

Info
General information about the API.

License
License information for the exposed API.

MediaType
Each Media Type Object provides schema and examples for the media type identified by its key.

Operation
Describes a single API operation on a path.

Parameter
Describes a single operation parameter. A unique parameter is defined by a combination of a name and location.

PasswordFlow
Configuration details for a password OAuth Flow See [link] [link](https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md#oauth-flow-object

PathItem
Describes the operations available on a single path.

RequestBody
Describes a single request body.

Response
Describes a single response from an API Operation, including design-time, static links to operations based on the response.

Schema
The Schema Object allows the definition of input and output data types. These types can be objects, but also primitives and arrays. This object is an extended subset of the JSON Schema Specification Wright Draft 00. For more information about the properties, see JSON Schema Core and JSON Schema Validation. Unless stated otherwise, the property definitions follow the JSON Schema.

Server
An object representing a Server.

ServerVariable
An object representing a Server Variable for server URL template substitution.

Spec
top level document

Tag
Adds metadata to a single tag that is used by the Operation Object. It is not mandatory to have a Tag Object per tag defined in the Operation Object instances.

Enums

BooleanObjectOrReference

Link
The Link object represents a possible design-time link for a response.

MediaTypeExample

ObjectOrReference

SecurityScheme
Defines a security scheme that can be used by the operations. Supported schemes are HTTP authentication, an API key (either as a header or as a query parameter), OAuth2's common flows (implicit, password, application and access code) as defined in RFC6749, and OpenID Connect Discovery.

Value
Represents any valid JSON value.